refactor: run selected linter init hooks Signed-off-by: Gregor Zeitlinger <gregor.zeitlinger@grafana.com>
