+++ src/init/hooks.rs
@@ -1,0 +1,6 @@
+pub(crate) mod biome;
+pub(crate) mod renovate_deps;
+pub(crate) mod rumdl;
+pub(crate) mod rustfmt;
+pub(crate) mod taplo;
+pub(crate) mod yamllint;
