+++ src/init/tests.rs
@@ -338,10 +338,10 @@ fn get_existing_config_dir_absent() {
 
 #[test]
 fn generate_rumdl_config_writes_file() {
-    use config_files::generate_rumdl_config;
+    use hooks::rumdl::generate_config;
     let tmp = tempfile::TempDir::new().unwrap();
     let config_dir = tmp.path().join(".github/config");
-    let written = generate_rumdl_config(tmp.path(), &config_dir, DEFAULT_LINE_LENGTH).unwrap();
+    let written = generate_config(tmp.path(), &config_dir, DEFAULT_LINE_LENGTH).unwrap();
     assert!(written);
     let content = std::fs::read_to_string(config_dir.join(".rumdl.toml")).unwrap();
     assert!(content.contains("line-length = 120"));
@@ -353,12 +353,12 @@ fn generate_rumdl_config_writes_file() {
 
 #[test]
 fn generate_rumdl_config_skips_when_target_exists() {
-    use config_files::generate_rumdl_config;
+    use hooks::rumdl::generate_config;
     let tmp = tempfile::TempDir::new().unwrap();
     let config_dir = tmp.path().join(".github/config");
     std::fs::create_dir_all(&config_dir).unwrap();
     std::fs::write(config_dir.join(".rumdl.toml"), "existing").unwrap();
-    let written = generate_rumdl_config(tmp.path(), &config_dir, DEFAULT_LINE_LENGTH).unwrap();
+    let written = generate_config(tmp.path(), &config_dir, DEFAULT_LINE_LENGTH).unwrap();
     assert!(!written);
     let content = std::fs::read_to_string(config_dir.join(".rumdl.toml")).unwrap();
     assert_eq!(content, "existing");
@@ -366,11 +366,11 @@ fn generate_rumdl_config_skips_when_target_exists() {
 
 #[test]
 fn generate_rumdl_config_replaces_legacy_json() {
-    use config_files::generate_rumdl_config;
+    use hooks::rumdl::generate_config;
     let tmp = tempfile::TempDir::new().unwrap();
     let config_dir = tmp.path().join(".github/config");
     std::fs::write(tmp.path().join(".markdownlint.json"), r#"{"MD013":false}"#).unwrap();
-    let written = generate_rumdl_config(tmp.path(), &config_dir, DEFAULT_LINE_LENGTH).unwrap();
+    let written = generate_config(tmp.path(), &config_dir, DEFAULT_LINE_LENGTH).unwrap();
     assert!(written);
     assert!(!tmp.path().join(".markdownlint.json").exists());
     let content = std::fs::read_to_string(config_dir.join(".rumdl.toml")).unwrap();
@@ -380,7 +380,7 @@ fn generate_rumdl_config_replaces_legacy_json() {
 
 #[test]
 fn generate_rumdl_config_converts_legacy_yaml() {
-    use config_files::generate_rumdl_config;
+    use hooks::rumdl::generate_config;
     let tmp = tempfile::TempDir::new().unwrap();
     let config_dir = tmp.path().join(".github/config");
     std::fs::write(
@@ -401,7 +401,7 @@ MD041: false
 "#,
     )
     .unwrap();
-    let written = generate_rumdl_config(tmp.path(), &config_dir, DEFAULT_LINE_LENGTH).unwrap();
+    let written = generate_config(tmp.path(), &config_dir, DEFAULT_LINE_LENGTH).unwrap();
     assert!(written);
     assert!(!tmp.path().join(".markdownlint.yaml").exists());
     let content = std::fs::read_to_string(config_dir.join(".rumdl.toml")).unwrap();
@@ -582,10 +582,10 @@ fn disable_editorconfig_line_length_for_patterns_is_idempotent() {
 
 #[test]
 fn generate_yamllint_config_writes_file() {
-    use config_files::generate_yamllint_config;
+    use hooks::yamllint::generate_config;
     let tmp = tempfile::TempDir::new().unwrap();
     let config_dir = tmp.path().join(".github/config");
-    let written = generate_yamllint_config(&config_dir, DEFAULT_LINE_LENGTH).unwrap();
+    let written = generate_config(&config_dir, DEFAULT_LINE_LENGTH).unwrap();
     assert!(written);
     let content = std::fs::read_to_string(config_dir.join(".yamllint.yml")).unwrap();
     assert_eq!(
@@ -596,10 +596,10 @@ fn generate_yamllint_config_writes_file() {
 
 #[test]
 fn generate_taplo_config_writes_file() {
-    use config_files::generate_taplo_config;
+    use hooks::taplo::generate_config;
     let tmp = tempfile::TempDir::new().unwrap();
     let config_dir = tmp.path().join(".github/config");
-    let written = generate_taplo_config(&config_dir, DEFAULT_LINE_LENGTH).unwrap();
+    let written = generate_config(&config_dir, DEFAULT_LINE_LENGTH).unwrap();
     assert!(written);
     let content = std::fs::read_to_string(config_dir.join(".taplo.toml")).unwrap();
     assert!(content.contains("[formatting]"));
@@ -609,12 +609,12 @@ fn generate_taplo_config_writes_file() {
 
 #[test]
 fn generate_taplo_config_skips_existing_supported_file() {
-    use config_files::generate_taplo_config;
+    use hooks::taplo::generate_config;
     let tmp = tempfile::TempDir::new().unwrap();
     let config_dir = tmp.path().join(".github/config");
     std::fs::create_dir_all(&config_dir).unwrap();
     std::fs::write(config_dir.join(".taplo.toml"), "existing").unwrap();
-    let written = generate_taplo_config(&config_dir, DEFAULT_LINE_LENGTH).unwrap();
+    let written = generate_config(&config_dir, DEFAULT_LINE_LENGTH).unwrap();
     assert!(!written);
     let content = std::fs::read_to_string(config_dir.join(".taplo.toml")).unwrap();
     assert_eq!(content, "existing");
@@ -622,22 +622,22 @@ fn generate_taplo_config_skips_existing_supported_file() {
 
 #[test]
 fn generate_taplo_config_skips_existing_legacy_name() {
-    use config_files::generate_taplo_config;
+    use hooks::taplo::generate_config;
     let tmp = tempfile::TempDir::new().unwrap();
     let config_dir = tmp.path().join(".github/config");
     std::fs::create_dir_all(&config_dir).unwrap();
     std::fs::write(config_dir.join("taplo.toml"), "existing").unwrap();
-    let written = generate_taplo_config(&config_dir, DEFAULT_LINE_LENGTH).unwrap();
+    let written = generate_config(&config_dir, DEFAULT_LINE_LENGTH).unwrap();
     assert!(!written);
     assert!(!config_dir.join(".taplo.toml").exists());
 }
 
 #[test]
 fn generate_rustfmt_config_writes_file() {
-    use config_files::generate_rustfmt_config;
+    use hooks::rustfmt::generate_config;
     let tmp = tempfile::TempDir::new().unwrap();
     let config_dir = tmp.path().join(".github/config");
-    let written = generate_rustfmt_config(&config_dir, DEFAULT_LINE_LENGTH).unwrap();
+    let written = generate_config(&config_dir, DEFAULT_LINE_LENGTH).unwrap();
     assert!(written);
     let content = std::fs::read_to_string(config_dir.join("rustfmt.toml")).unwrap();
     assert_eq!(content, "max_width = 120\n");
@@ -645,12 +645,12 @@ fn generate_rustfmt_config_writes_file() {
 
 #[test]
 fn generate_rustfmt_config_skips_existing_file() {
-    use config_files::generate_rustfmt_config;
+    use hooks::rustfmt::generate_config;
     let tmp = tempfile::TempDir::new().unwrap();
     let config_dir = tmp.path().join(".github/config");
     std::fs::create_dir_all(&config_dir).unwrap();
     std::fs::write(config_dir.join("rustfmt.toml"), "existing").unwrap();
-    let written = generate_rustfmt_config(&config_dir, DEFAULT_LINE_LENGTH).unwrap();
+    let written = generate_config(&config_dir, DEFAULT_LINE_LENGTH).unwrap();
     assert!(!written);
     let content = std::fs::read_to_string(config_dir.join("rustfmt.toml")).unwrap();
     assert_eq!(content, "existing");
@@ -658,9 +658,9 @@ fn generate_rustfmt_config_skips_existing_file() {
 
 #[test]
 fn generate_biome_config_writes_file() {
-    use config_files::generate_biome_config;
+    use hooks::biome::generate_config;
     let tmp = tempfile::TempDir::new().unwrap();
-    let written = generate_biome_config(tmp.path()).unwrap();
+    let written = generate_config(tmp.path()).unwrap();
     assert!(written);
     let content = std::fs::read_to_string(tmp.path().join("biome.jsonc")).unwrap();
     assert!(content.contains("\"indentStyle\": \"space\""));
@@ -669,10 +669,10 @@ fn generate_biome_config_writes_file() {
 
 #[test]
 fn generate_biome_config_skips_existing_jsonc() {
-    use config_files::generate_biome_config;
+    use hooks::biome::generate_config;
     let tmp = tempfile::TempDir::new().unwrap();
     std::fs::write(tmp.path().join("biome.jsonc"), "existing").unwrap();
-    let written = generate_biome_config(tmp.path()).unwrap();
+    let written = generate_config(tmp.path()).unwrap();
     assert!(!written);
     let content = std::fs::read_to_string(tmp.path().join("biome.jsonc")).unwrap();
     assert_eq!(content, "existing");
@@ -680,10 +680,10 @@ fn generate_biome_config_skips_existing_jsonc() {
 
 #[test]
 fn generate_biome_config_migrates_legacy_supported_json_name() {
-    use config_files::generate_biome_config;
+    use hooks::biome::generate_config;
     let tmp = tempfile::TempDir::new().unwrap();
     std::fs::write(tmp.path().join("biome.json"), "existing").unwrap();
-    let written = generate_biome_config(tmp.path()).unwrap();
+    let written = generate_config(tmp.path()).unwrap();
     assert!(written);
     assert!(!tmp.path().join("biome.json").exists());
     let content = std::fs::read_to_string(tmp.path().join("biome.jsonc")).unwrap();
@@ -694,14 +694,8 @@ fn generate_biome_config_migrates_legacy_supported_json_name() {
 fn generate_flint_toml_writes_skeleton() {
     let tmp = tempfile::TempDir::new().unwrap();
     let dir = tmp.path().join("config");
-    let written = generate_flint_toml(
-        &dir,
-        "main",
-        crate::setup::V2_BASELINE_SETUP_VERSION,
-        false,
-        None,
-    )
-    .unwrap();
+    let written =
+        generate_flint_toml(&dir, "main", crate::setup::V2_BASELINE_SETUP_VERSION).unwrap();
     assert!(written);
     let content = std::fs::read_to_string(dir.join("flint.toml")).unwrap();
     assert!(content.contains("[settings]"));
@@ -716,8 +710,6 @@ fn generate_flint_toml_non_main_branch() {
         tmp.path(),
         "master",
         crate::setup::V2_BASELINE_SETUP_VERSION,
-        false,
-        None,
     )
     .unwrap();
     assert!(written);
@@ -726,115 +718,11 @@ fn generate_flint_toml_non_main_branch() {
 }
 
 #[test]
-fn generate_flint_toml_with_renovate_placeholder() {
-    let tmp = tempfile::TempDir::new().unwrap();
-    generate_flint_toml(
-        tmp.path(),
-        "main",
-        crate::setup::V2_BASELINE_SETUP_VERSION,
-        true,
-        None,
-    )
-    .unwrap();
-    let content = std::fs::read_to_string(tmp.path().join("flint.toml")).unwrap();
-    assert!(content.contains("[checks.renovate-deps]"));
-    assert!(content.contains("# exclude_managers ="));
-}
-
-#[test]
-fn generate_flint_toml_with_renovate_managers() {
-    let tmp = tempfile::TempDir::new().unwrap();
-    let managers = vec!["github-actions".to_string(), "cargo".to_string()];
-    generate_flint_toml(
-        tmp.path(),
-        "main",
-        crate::setup::V2_BASELINE_SETUP_VERSION,
-        true,
-        Some(&managers),
-    )
-    .unwrap();
-    let content = std::fs::read_to_string(tmp.path().join("flint.toml")).unwrap();
-    assert!(content.contains("[checks.renovate-deps]"));
-    assert!(
-        content.contains("exclude_managers = [\"github-actions\", \"cargo\"]"),
-        "managers written uncommented: {content}"
-    );
-    assert!(!content.contains("# exclude_managers"));
-}
-
-#[test]
-fn generate_flint_toml_keeps_existing_without_migrated_renovate_managers() {
-    let tmp = tempfile::TempDir::new().unwrap();
-    std::fs::write(tmp.path().join("flint.toml"), "[settings]\n").unwrap();
-    let written = generate_flint_toml(
-        tmp.path(),
-        "main",
-        crate::setup::V2_BASELINE_SETUP_VERSION,
-        true,
-        None,
-    )
-    .unwrap();
-    assert!(!written);
-    let content = std::fs::read_to_string(tmp.path().join("flint.toml")).unwrap();
-    assert_eq!(content, "[settings]\n");
-}
-
-#[test]
-fn generate_flint_toml_patches_existing_with_renovate_managers() {
-    let tmp = tempfile::TempDir::new().unwrap();
-    std::fs::write(tmp.path().join("flint.toml"), "[settings]\n").unwrap();
-    let managers = vec!["github-actions".to_string(), "cargo".to_string()];
-    let written = generate_flint_toml(
-        tmp.path(),
-        "main",
-        crate::setup::V2_BASELINE_SETUP_VERSION,
-        true,
-        Some(&managers),
-    )
-    .unwrap();
-    assert!(written);
-    let content = std::fs::read_to_string(tmp.path().join("flint.toml")).unwrap();
-    assert!(
-        content.contains("exclude_managers = [\"github-actions\", \"cargo\"]"),
-        "managers written uncommented: {content}"
-    );
-}
-
-#[test]
-fn generate_flint_toml_keeps_existing_renovate_managers() {
-    let tmp = tempfile::TempDir::new().unwrap();
-    std::fs::write(
-        tmp.path().join("flint.toml"),
-        "[checks.renovate-deps]\nexclude_managers = [\"npm\"]\n",
-    )
-    .unwrap();
-    let managers = vec!["github-actions".to_string(), "cargo".to_string()];
-    let written = generate_flint_toml(
-        tmp.path(),
-        "main",
-        crate::setup::V2_BASELINE_SETUP_VERSION,
-        true,
-        Some(&managers),
-    )
-    .unwrap();
-    assert!(!written);
-    let content = std::fs::read_to_string(tmp.path().join("flint.toml")).unwrap();
-    assert!(content.contains("exclude_managers = [\"npm\"]"));
-    assert!(!content.contains("github-actions"));
-}
-
-#[test]
 fn generate_flint_toml_skips_existing() {
     let tmp = tempfile::TempDir::new().unwrap();
     std::fs::write(tmp.path().join("flint.toml"), "existing content").unwrap();
-    let written = generate_flint_toml(
-        tmp.path(),
-        "main",
-        crate::setup::V2_BASELINE_SETUP_VERSION,
-        false,
-        None,
-    )
-    .unwrap();
+    let written =
+        generate_flint_toml(tmp.path(), "main", crate::setup::V2_BASELINE_SETUP_VERSION).unwrap();
     assert!(!written);
     let content = std::fs::read_to_string(tmp.path().join("flint.toml")).unwrap();
     assert_eq!(content, "existing content");
